+++ src/main/java/uk/gov/hmcts/reform/adoption/adoptioncase/model/ManageOrdersDataAdditionalParagraph.java
@@ -12,7 +12,7 @@ import static uk.gov.hmcts.ccd.sdk.type.FieldType.TextArea;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class AdditionalParagraph {
+public class ManageOrdersDataAdditionalParagraph {
     @CCD(label = "Copy and paste or type any additional orders directly into the text box. "
         + "You can include any orders for step parents, for example, or inclusion of other parties to attend.",
         typeOverride = TextArea)
